Refactor argument parser
The previous version was a bit too clever with the way it used types,
requiring the user to consume an option before getting the argument
iterator back. This was intended to force the user to indicate whether
an optional argument accepted a parameter, so we could correctly parse
"-flag value" as either two arguments or as a single "-flag=value"
argument.

In general there seems to be a tension between how often we catch usage
errors with the type system, and how easy the library is to use. This
refactoring attempts to strike a better balance.

Additional notes:

- We can now work with String or OsString equally. The types are a bit
  clumsy, since we use `<T as Iterator>::Item` everywhere.

- Errors from parsing optional values are now boxed and passed out.

- Custom errors are gone. This might be a mistake!

- Much better documentation.